add method to get cause
